Update include locations since moving to Utilities, and needed to add a few libraries to make it build. Not sure the lardata one is the minimal one, but can be edited later if needed.
